==> notebooks/04_model_deployment.ipynb <==
# **Model Deployment & Inference**

After training a model and evaluate its performance, the next steps are:
<ul>
<li>Load the piepline back</li>
<li>Make new predictions</li>
<li>Simulate real world usage</li>
</ul>


### **1. Imports**

import joblib
from sklearn.pipeline import Pipeline
import pandas as pd

### **2. Load the Pipeline**

loaded_model=joblib.load("../models/revenue_prediction_pipeline.pkl")
print("model loaded successfully")

### **3. Simulate New Business Data** 

new_data = pd.DataFrame({
    "marketing_spend": [5000],
    "website_visits": [20000],
    "conversion_rate": [0.05],
    "num_customers": [1000],
    "avg_order_value": [50],
    "discount_rate": [0.10],
    "season": ["Summer"]
})

prediction = loaded_model.predict(new_data)

print("Predicted Revenue : ", prediction[0])

This prediction can now be used by : 
<ul>
<li>Small business owner</li>
<li>Financial planner</li>
<li>Growth Strategist</li>
</ul>
